refactor: 아이템 서비스

+++ item_service/app/api/categories.py
@@ -37,7 +37,6 @@ def post_category() :
     store_id = get_jwt_identity()
 
     req = request.get_json()
-    print(req['name'])
 
     new_category = Categories(
         name=req['name'],
@@ -82,6 +81,14 @@ def put_category(category_id: int) :
 # 카테고리 삭제
 @bp.route('/<int:category_id>', methods=['DELETE'])
 def delete_category(category_id: int) :
+
+    category = Categories.query.filter_by(category_id=category_id).first()
+
+    if category is None:
+        return {
+            "result": "failed",
+            "message": '카테고리가 존재하지 않습니다.'
+        }, 404
 
     db.session.query(Categories).filter_by(category_id=category_id).delete()
